+++ orthoseg/model/model_factory.py
@@ -27,7 +27,6 @@ if KERAS_GTE_3:
 else:
     from keras import backend as ops
 
-
 # Get a logger...
 logger = logging.getLogger(__name__)
 
@@ -595,6 +594,24 @@ def load_model(
     # model.run_eagerly = True
 
     return model, preprocess_input_func
+
+
+def set_dtype_policy(policy: str):
+    """Set the keras dtype policy.
+
+    Args:
+        policy (str): the policy to set. One of the following options:
+            - "mixed_float16": Use mixed precision with float16 for computations and
+              float32 for variables.
+            - "float32": Use float32 for computations and variables.
+            - "float64": Use float64 for computations and variables.
+            - "mixed_bfloat16": Use mixed precision with bfloat16 for computations and
+              float32 for variables.
+    """
+    if KERAS_GTE_3:
+        keras.config.set_dtype_policy(policy)
+    else:
+        keras.mixed_precision.set_global_policy(policy)
 
 
 def set_trainable(model, recompile: bool = True):